+++ index.js
@@ -1,0 +1,6 @@
+/**
+ * index
+ * @module www
+ */
+
+const www = require('./bin/www');